# openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /task:
        get:
            tags:
                - TaskCURD
            operationId: TaskCURD_GetTaskList
            parameters:
                - name: IDList
                  in: query
                  schema:
                    type: array
                    items:
                        type: integer
                        format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.GetTaskListResponse'
        post:
            tags:
                - TaskCURD
            description: |-
                --------------------------------------------------
                 tbl : task
            operationId: TaskCURD_AddTask
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.AddTaskRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.AddTaskResponse'
        delete:
            tags:
                - TaskCURD
            operationId: TaskCURD_DelTaskByIDList
            parameters:
                - name: IDList
                  in: query
                  schema:
                    type: array
                    items:
                        type: integer
                        format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.Empty'
        patch:
            tags:
                - TaskCURD
            operationId: TaskCURD_UpdateTask
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.UpdateTaskRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.UpdateTaskResponse'
...
components:
    schemas:
...
        api.AddTaskRequest:
            type: object
            properties:
                task:
                    $ref: '#/components/schemas/api.TaskInfo'
        api.AddTaskResponse:
            type: object
            properties:
                task:
                    $ref: '#/components/schemas/api.TaskInfo'
...
        api.Empty:
            type: object
            properties: {}
...
        api.GetTaskListResponse:
            type: object
            properties:
                taskList:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.TaskInfo'
...
        api.TaskInfo:
            type: object
            properties:
                ID:
                    type: integer
                    format: int32
                parentID:
                    type: integer
                    format: int32
                prevID:
                    type: integer
                    format: int32
                task:
                    type: string
                status:
                    type: integer
                    format: int32
                estimate:
                    type: integer
                    format: int32
                start:
                    type: string
                end:
                    type: string
                desc:
                    type: string
                metadata:
                    type: string
                createTime:
                    type: string
            description: |-
                --------------------------------------------------
                 tbl : task
...
        api.UpdateTaskRequest:
            type: object
            properties:
                task:
                    $ref: '#/components/schemas/api.TaskInfo'
        api.UpdateTaskResponse:
            type: object
            properties:
                task:
                    $ref: '#/components/schemas/api.TaskInfo'
...
servers:
    - description: PGO API
      url: http://127.0.0.1:8080